Developing token issue and validation
